requirements.txt
ansible==2.8.6
APScheduler==3.6.1
cffi==1.13.1
cryptography==2.8
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.2.6
django-import-export==1.2.0
djangorestframework==3.10.3
et-xmlfile==1.0.1
gunicorn==19.9.0
jdcal==1.4.1
Jinja2==2.10.3
MarkupPy==1.14
MarkupSafe==1.1.1
odfpy==1.4.0
openpyxl==3.0.0
Pillow==6.2.1
psycopg2==2.8.3
pycparser==2.19
pytz==2019.3
PyYAML==5.1.2
six==1.12.0
sqlparse==0.3.0
tablib==0.14.0
tzlocal==2.0.0
xlrd==1.2.0
xlwt==1.3.0
